api/requirements.txt
boto3==1.17.86
botocore==1.20.86
Flask==1.1.4
python-dotenv==0.17.1
requests==2.25.1
simplejson~=3.17.2
pydantic~=1.8.2